Restructure for ng packagr
