--- src/app/app.component.html
@@ -1,64 +1,0 @@
-<h2>Stand-alone</h2>
-<label>Date & Time Picker</label>
-<ngx-datetime-picker [(selectedDateTime)]="dateTimeExample"></ngx-datetime-picker>
-<ngx-datetime-picker [(selectedDateTime)]="dateTimeExample" min="6/1/2018"></ngx-datetime-picker>
-<ngx-datetime-picker [(selectedDateTime)]="dateTimeExample" max="6/1/2018"></ngx-datetime-picker>
-<ngx-datetime-picker [(selectedDateTime)]="dateTimeExample" [disableButton]="true"></ngx-datetime-picker>
-<ngx-datetime-picker [(selectedDateTime)]="dateTimeExample" [disableInput]="true"></ngx-datetime-picker>
-<ngx-datetime-picker [(selectedDateTime)]="dateTimeExample" [disablePicker]="true"></ngx-datetime-picker>
-<ngx-datetime-picker [(selectedDateTime)]="dateTimeExample" [doNotCloseOnDateSet]="true"></ngx-datetime-picker>
-<ngx-datetime-picker [(selectedDateTime)]="dateTimeExample" [use24HourClock]="true"></ngx-datetime-picker>
-<ngx-datetime-picker [(selectedDateTime)]="dateTimeExample" [styles]="{ input: { width: '100%' } , button: { 'margin-left': '5px' }}"></ngx-datetime-picker>
-<hr />
-<label>Date Picker</label>
-<ngx-date-picker [(selectedDate)]="dateExample" #pickerToFocus></ngx-date-picker>
-<ngx-date-picker [(selectedDate)]="dateExample" [inputTabIndex]="1"></ngx-date-picker>
-<ngx-date-picker [(selectedDate)]="dateExample" min="6/1/2018"></ngx-date-picker>
-<ngx-date-picker [(selectedDate)]="dateExample" max="6/1/2018"></ngx-date-picker>
-<ngx-date-picker [(selectedDate)]="dateExample" min="6/1/2018" max="8/5/2019"></ngx-date-picker>
-<ngx-date-picker [(selectedDate)]="dateExample" [disableButton]="true"></ngx-date-picker>
-<ngx-date-picker [(selectedDate)]="dateExample" [disableInput]="true"></ngx-date-picker>
-<ngx-date-picker [(selectedDate)]="dateExample" [disablePicker]="true"></ngx-date-picker>
-<ngx-date-picker [(selectedDate)]="dateExample" [doNotCloseOnDateSet]="true"></ngx-date-picker>
-<hr />
-<label>Time Picker</label>
-<ngx-time-picker [(selectedTime)]="timeExample"></ngx-time-picker>
-<ngx-time-picker [(selectedTime)]="timeExample" [disableButton]="true"></ngx-time-picker>
-<ngx-time-picker [(selectedTime)]="timeExample" [disableInput]="true"></ngx-time-picker>
-<ngx-time-picker [(selectedTime)]="timeExample" [disablePicker]="true"></ngx-time-picker>
-<ngx-time-picker [(selectedTime)]="timeExample" [use24HourClock]="true"></ngx-time-picker>
-
-<br>
-<button (click)="dateExample=null;timeExample=null;dateTimeExample=null;" >Clear All</button>
-<button (click)="setToNow()">Set all to now</button>
-<button (click)="focusPicker()">Focus datepicker</button>
-
-<br>
-
-{{ dateTimeExample }}
-<br>
-{{ dateExample }}
-<br>
-{{ timeExample }}
-
-<h2>With angular forms</h2>
-<label>Date & Time Picker</label>
-<ngx-datetime-picker [(ngModel)]="dateTimeExample"></ngx-datetime-picker>
-<hr />
-<label>Date Picker</label>
-<ngx-date-picker [(ngModel)]="dateExample"></ngx-date-picker>
-<hr />
-<label>Time Picker</label>
-<ngx-time-picker [(ngModel)]="timeExample"></ngx-time-picker>
-
-<br>
-<button type="button" (click)="dateExample=null;timeExample=null;dateTimeExample=null;" >Clear All</button>
-<button type="button" (click)="setToNow()">Set all to now</button>
-
-<br>
-
-{{ dateTimeExample }}
-<br>
-{{ dateExample }}
-<br>
-{{ timeExample }}
