--- src/app/index.ts
@@ -1,4 +1,0 @@
-export { DateTimePickerModule } from './datetimepicker.module';
-export { TimePickerComponent } from './timePicker.component/timePicker.component';
-export { DatePickerComponent } from './datePicker.component/datePicker.component';
-export { DateTimePickerComponent } from './dateTimePicker.component/dateTimePicker.component';
