+++ src/components/datePicker/datePicker.component.ts
@@ -1,9 +1,9 @@
 import {Component, OnInit, Input, Output, EventEmitter, ViewEncapsulation, HostListener, ElementRef, forwardRef, ViewChild} from '@angular/core';
 import {ControlValueAccessor, NG_VALUE_ACCESSOR} from '@angular/forms';
-import {IsMobileService} from '../services/isMobile.service';
-import {DateService} from '../services/date.service';
-import {StyleObject} from '../models/styleObject.model';
-import {Renderer} from '../services/renderer.service';
+import {IsMobileService} from '../../services/isMobile.service';
+import {DateService} from '../../services/date.service';
+import {StyleObject} from '../../models/styleObject.model';
+import {Renderer} from '../../services/renderer.service';
 
 @Component({
 	selector: 'ngx-date-picker',
@@ -15,9 +15,8 @@ import {Renderer} from '../services/renderer.service';
 			useExisting: forwardRef(() => DatePickerComponent),
 			multi: true,
 		},
-	],
+	]
 })
-
 export class DatePickerComponent implements OnInit, ControlValueAccessor {
     @Input() selectedDate: Date;
     @Input() min: string;
