--- src/index.html
@@ -1,14 +1,0 @@
-<!doctype html>
-<html>
-<head>
-  <meta charset="utf-8">
-  <title>NgxDatetimepicker</title>
-  <base href="/">
-
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-  <link rel="icon" type="image/x-icon" href="favicon.ico">
-</head>
-<body>
-  <app-root>Loading...</app-root>
-</body>
-</html>
