+++ src/models/styleObject.model.ts
