+++ src/scss/_picker.scss
