--- src/tsconfig.json
@@ -1,18 +1,0 @@
-{
-  "compilerOptions": {
-    "baseUrl": "tsconfig",
-    "declaration": true,
-    "emitDecoratorMetadata": true,
-    "experimentalDecorators": true,
-    "lib": ["es6", "dom"],
-    "mapRoot": "./",
-    "module": "es6",
-    "moduleResolution": "node",
-    "sourceMap": true,
-    "target": "es5",
-    "typeRoots": [
-      "../node_modules/@types",
-        "./typings/isMobile.d.ts"
-    ]
-  }
-}
